refactor: extract shared wasmtime runtime module in wado-cli
- Create runtime.rs with shared wasmtime configuration
- WasiState, create_config, create_engine, create_store, create_linker
- Disable Cranelift optimizations for faster dev compilation
- Simplify run.rs and test.rs by using shared runtime module
